Added onSafe event (UNTESTED)

+++ CISC-226-GAME/Assets/Scripts/AnimalScripts/MovementSM.cs
@@ -26,7 +26,7 @@ public class MovementSM : StateMachine
     [HideInInspector]
     public BoxCollider2D collider2D;
 
-    public float id;
+    public int id;
 
     // Hide HideInInspector makes the variables only accessable in the script
     // [HideInInspector]
